Add a new tracing option, AVR_MMCU_TAG_VCD_IO_IRQ, that allows any Input/Output IRQ to be logged in a VCD file.  Make it available in run_avr.  Also add a new macro, AVR_MCU_VCD_REGISTER_BIT(), to simplify the tracing of peripheral register bits. Use these in atmega88_example.c, fixing the broken trace file reported in Bug #486 - "Unable to reproduce tests/atmega88_example.axf results" as an alternative to PR #547 - "bugfix: Fix Test VCD Output".

+++ simavr/sim/sim_elf.h
@@ -69,7 +69,7 @@ typedef struct elf_firmware_t {
 	struct {
 		uint8_t kind;
 		uint8_t mask;
-		uint16_t addr;
+		uint32_t addr; // May be an ioctl value.
 		char	name[64];
 	} trace[32];
